sdpx: Avoid loading of SPDX_LICENSE_DATA into global config
Loading a load of json files into a memory structure and stashing in a bitbake
variable is relatively anti-social making bitbake -e output hard to read for
example as well as other potential performance issues.

Defer loading of that data until it is actually needed/used in a funciton
where it is now passed as a parameter.

(From OE-Core rev: 8a80e71042b15ba1cd3b8827b6fab48265b1e692)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>