kernel-fit-image.bbclass: add a new FIT image implementation
This new recipe and the new kernel-fit-image.bbclass are intended to
become the successor of the kernel-fitimage.bbclass.

This recipe is an alternative implementation of what the
kernel-fitimage.bbclass provides. Instead of injecting the FIT image
related build steps into the kernel recipe, the new recipe takes the
kernel artifacts from the kernel recipe and creates the FIT image as a
independent task.

This solves some basic problems:
* sstate does not work well when a fitImage contains an initramfs. The
  kernel is rebuilt from scratch if the build runs from an empty TMPDIR.
* A fitImage kernel is not available as a package, but all other kernel
  image types are.
* The task dependencies in the kernel are very complex and difficult to
  debug if something goes wrong. As a separate, downstream recipe, this
  is now much easier.

The long storry about this issue is here:
[YOCTO #12912]

(From OE-Core rev: 5b02c4cb3f1e52431723b494ffcf599aaf4a6f2b)

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>