kernel.bbclass: remove support for type fitImage
Support for fitImage inf KERNEL_IMAGETYPE or KERNEL_IMAGETYPES has been
removed from kernel.bbclass. There is now the new
kernel-fit-image.bbclass which allows to write an independent recipe
like the linux-yocto-fitimage.bb which provides the same functionality.
This new recipe has the advantage that it can use kernel artifacts from
the sstate-cache when the initramfs changes but the kernel remains
unchanged. This was not possible with the old kernel-fitimage.bbclass.

The combination of INITRAMFS_IMAGE_BUNDLE = “1” and the creation of a
FIT image is no longer supported. But there is probably no real use case
for such a configuration. The FIT image can contain multiple artifacts
such as a kernel and an initramfs without bundling the initramfs into
the kernel artifact. Bundling the initramfs into the kernel artifact
also requires the full kernel build directory, which inherently prevents
the use of the kernel from the sstate cache.

(From OE-Core rev: 1d728e46d13da99fa3987f032fb34461eb0392e3)

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>