oe-selftest: fitimage: drop initramfs bundle test
This test case is no longer useful. Sure, it is possible to build a
kernel with an initramfs bundled. But since this is not the kernel
which gets included into the FIT image, there is no need for testing
this as part of the FIT image test suite.

(From OE-Core rev: f4a57b7e48c41f7ecf656837372e6878580364fc)

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>